Get loadRecords working on page load
